project_resample_daps/frozen_requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.4.0
backports.zstd @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_backports.zstd_1764345088/work
beautifulsoup4==4.14.3
blinker==1.9.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1764016952863/work
cachetools==6.2.2
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.7
-e git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1#egg=clip
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy==1.0.5
cycler==0.11.0
diffusers==0.32.2
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.14.0
fonttools==4.37.2
frozenlist==1.8.0
fsspec==2024.5.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1762946744175/work
grpcio==1.74.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
huggingface-hub==0.29.3
hydra-core==1.3.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna==3.4
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.4
kornia==0.7.3
kornia_rs==0.1.10
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
lightning-utilities==0.15.2
Markdown==3.10
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict==6.7.0
munkres==1.1.4
narwhals==2.13.0
networkx==3.1
numpy==1.24.4
omegaconf==2.3.0
opencv-python-headless==4.7.0.72
packaging==21.3
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1764614464206/work
pillow==10.3.0
piq==0.8.0
platformdirs==4.2.2
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1733239724146/work
prettytable==3.10.0
propcache==0.4.1
protobuf==4.25.3
psutil==5.9.8
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.0.9
PyQt5==5.15.11
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.0.8
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyWavelets==1.8.0
PyYAML==6.0
referencing==0.37.0
regex==2025.11.3
requests==2.28.1
rpds-py==0.30.0
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==2.2.1
setproctitle==1.3.3
sip @ file:///home/conda/feedstock_root/build_artifacts/sip_1759437834046/work
six==1.16.0
smmap==5.0.1
soupsieve==2.8
streamlit==1.52.1
sympy==1.12
taming-transformers-rom1504==0.0.6
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
test_tube==0.7.5
tifffile @ file:///croot/tifffile_1734340907867/work
tokenizers==0.20.3
toml @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_toml_1764486833/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1760014679/work
torch==2.3.0
torchaudio==2.3.0
torchdiffeq==0.2.5
torchmetrics==1.8.2
torchvision==0.18.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1762506714284/work
tqdm==4.64.1
transformers==4.46.3
triton==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1763054696114/work
urllib3==1.26.12
wandb==0.17.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.4
yarl==1.22.0
zipp==3.20.2
